+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## current
 - Ask for database to load on start-up
 - Add page break to report file
+- Implemented player/goalie deletion
 
 ## 0.14.0
 - Fixed a bug that was causing shutouts to not be recorded